feat: single scenario view
