+++ src/features/plays/pages/play/components/scenario-modal.tsx
@@ -1,7 +1,3 @@
-import { useEffect } from 'react'
-import * as z from 'zod'
-import { useForm } from 'react-hook-form'
-import { zodResolver } from '@hookform/resolvers/zod'
 import { Trash2 } from 'lucide-react'
 import { toast } from 'sonner'
 import { Button } from '@/components/ui/button'
@@ -14,31 +10,9 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog'
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@/components/ui/form'
-import { Input } from '@/components/ui/input'
-import { Textarea } from '@/components/ui/textarea'
 import Confirmable from '@/components/confirmable'
-import {
-  usePlaybookScenarioUpdateMutation,
-  usePlaybookScenarioCreateMutation,
-  usePlaybookScenariosDeleteMutation,
-} from '../../../graphql/operations.generated'
-
-const scenarioFormSchema = z.object({
-  name: z.string().min(1, 'Scenario name is required').trim(),
-  description: z.string().min(1, 'Description is required').trim(),
-  actionPlan: z.string().min(1, 'Action plan is required').trim(),
-})
-
-type ScenarioFormData = z.infer<typeof scenarioFormSchema>
+import { usePlaybookScenariosDeleteMutation } from '../../../graphql/operations.generated'
+import { ScenarioForm } from './scenario-form'
 
 interface ScenarioModalProps {
   mode: 'create' | 'edit'
@@ -62,70 +36,11 @@ export function ScenarioModal({
   onClose,
   onSuccess,
 }: ScenarioModalProps) {
-  const form = useForm<ScenarioFormData>({
-    resolver: zodResolver(scenarioFormSchema),
-    defaultValues: {
-      name: '',
-      description: '',
-      actionPlan: '',
-    },
-  })
-
-  const [updateScenario, { loading: updating }] = usePlaybookScenarioUpdateMutation()
-  const [createScenario, { loading: creating }] = usePlaybookScenarioCreateMutation()
   const [deleteScenarios, { loading: deleting }] = usePlaybookScenariosDeleteMutation()
 
-  const loading = updating || creating || deleting || form.formState.isSubmitting
-
-  useEffect(() => {
-    if (mode === 'edit' && scenario) {
-      form.reset({
-        name: scenario.name || '',
-        description: scenario.description || '',
-        actionPlan: scenario.actionPlan || '',
-      })
-    } else if (mode === 'create') {
-      form.reset({
-        name: '',
-        description: '',
-        actionPlan: '',
-      })
-    }
-  }, [scenario, mode, isOpen, form])
-
-  const onSubmit = async (data: ScenarioFormData) => {
-    try {
-      if (mode === 'edit' && scenario) {
-        await updateScenario({
-          variables: {
-            input: {
-              id: scenario.id,
-              name: data.name,
-              description: data.description,
-              actionPlan: data.actionPlan,
-            },
-          },
-        })
-        toast.success('Scenario updated successfully')
-      } else if (mode === 'create' && playbookId) {
-        await createScenario({
-          variables: {
-            input: {
-              playbookId,
-              name: data.name,
-              description: data.description,
-              actionPlan: data.actionPlan,
-            },
-          },
-        })
-        toast.success('Scenario created successfully')
-      }
-
-      onSuccess()
-      handleClose()
-    } catch (error: any) {
-      toast.error(`Failed to ${mode} scenario: ${error.message}`)
-    }
+  const handleFormSuccess = () => {
+    onSuccess()
+    handleClose()
   }
 
   const handleDelete = async () => {
@@ -165,79 +80,15 @@ export function ScenarioModal({
         </DialogHeader>
 
         <DialogBody>
-          <Form {...form}>
-            <form
-              id='scenario-form'
-              onSubmit={form.handleSubmit(onSubmit)}
-              className='space-y-6 py-4'
-            >
-              <FormField
-                control={form.control}
-                name='name'
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Name</FormLabel>
-                    <FormControl>
-                      <Input
-                        autoComplete='off'
-                        placeholder='Scenario name'
-                        disabled={loading}
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormDescription>Give your scenario a clear, descriptive name</FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name='description'
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Condition</FormLabel>
-                    <FormControl>
-                      <Textarea
-                        placeholder='Describe the condition that triggers this scenario'
-                        rows={4}
-                        disabled={loading}
-                        className='resize-none'
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormDescription>
-                      Explain in what situations this scenario applies
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name='actionPlan'
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Action Plan</FormLabel>
-                    <FormControl>
-                      <Textarea
-                        placeholder='Define the action plan for this scenario'
-                        rows={6}
-                        disabled={loading}
-                        className='resize-none'
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormDescription>
-                      Provide detailed steps that will be executed when this scenario runs
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </form>
-          </Form>
+          <ScenarioForm
+            mode={mode}
+            playbookId={playbookId}
+            scenario={scenario}
+            onSuccess={handleFormSuccess}
+            formId='scenario-form'
+            showButtons={false}
+            disabled={deleting}
+          />
         </DialogBody>
 
         <DialogFooter className='flex items-center justify-between'>
@@ -251,7 +102,7 @@ export function ScenarioModal({
                 variant='danger'
                 isLoading={deleting}
               >
-                <Button variant='destructive' size='sm' disabled={loading}>
+                <Button variant='destructive' size='sm' disabled={deleting}>
                   <Trash2 className='mr-2 h-4 w-4' />
                   Delete
                 </Button>
@@ -260,10 +111,10 @@ export function ScenarioModal({
           </div>
 
           <div className='flex gap-2'>
-            <Button variant='outline' onClick={handleClose} disabled={loading}>
+            <Button variant='outline' onClick={handleClose} disabled={deleting}>
               Cancel
             </Button>
-            <Button type='submit' form='scenario-form' loading={loading}>
+            <Button type='submit' form='scenario-form' disabled={deleting}>
               {mode === 'edit' ? 'Save Changes' : 'Create Scenario'}
             </Button>
           </div>
